# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Management API",
        "description": "API documentation for the Order Management System",
        "contact": {
            "email": "niyonshutivalentin@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/order-items": {
            "post": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Create an order item",
                "operationId": "1fa1e1f5c18b19fdaba4e08c27d16a04",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderItem"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderItemResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order-items/{id}": {
            "get": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Get a specific order item",
                "operationId": "8db93e729a0f278f71007a7d4cfd846b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderItemResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Update an order item",
                "operationId": "4feea364a5d4525f8aacfbae57eb35e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderItemResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Delete an order item",
                "operationId": "cef28e5ba5671e27cb5e871bd8db0755",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderItemResource": {
                "title": "Order Item Resource",
                "description": "Order Item Resource Response",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/OrderItem"
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "required": [
                    "id",
                    "order_id",
                    "product_id",
                    "quantity",
                    "price",
                    "total"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "description": "Foreign key referencing products table",
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 2
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 99.99
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 199.98
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "required": [
                    "id",
                    "name",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Laptop"
                    },
                    "description": {
                        "type": "string",
                        "example": "High-performance laptop"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}